[snap.auxil.execute] remove par from node if failed with right message
if certain node parameters are not supported by the used version of GPT
they are removed from the XML file, which is then passed to execute
again

+++ pyroSAR/snap/auxil.py
@@ -73,7 +73,7 @@ def parse_node(name):
     return Node(element)
 
 
-def execute(xmlfile, cleanup=True, gpt_exceptions=None):
+def execute(xmlfile, cleanup=True, gpt_exceptions=None, verbose=True):
     """
     execute SNAP workflows via the Graph Processing Tool gpt.
     This function merely calls gpt with some additional command
@@ -92,6 +92,8 @@ def execute(xmlfile, cleanup=True, gpt_exceptions=None):
         each (sub-)workflow containing this operator will be executed with the define executable;
         
          - e.g. ``{'Terrain-Flattening': '/home/user/snap/bin/gpt'}``
+    verbose: bool
+        print out status messages?
     
     Returns
     -------
@@ -115,7 +117,8 @@ def execute(xmlfile, cleanup=True, gpt_exceptions=None):
                 gpt_exec = exec
                 message += ' (using {})'.format(exec)
                 break
-    print(message)
+    if verbose:
+        print(message)
     # try to find the GPT executable
     if gpt_exec is None:
         try:
@@ -141,16 +144,29 @@ def execute(xmlfile, cleanup=True, gpt_exceptions=None):
     err = err.decode('utf-8') if isinstance(err, bytes) else err
     # delete intermediate files if an error occurred
     if proc.returncode != 0:
-        if cleanup:
-            if os.path.isfile(outname + '.tif'):
-                os.remove(outname + '.tif')
-            elif os.path.isdir(outname):
-                shutil.rmtree(outname)
-        print(out + err)
-        print('failed: {}'.format(os.path.basename(infile)))
-        err_match = re.search('Error: (.*)\n', out + err)
-        errmessage = err_match.group(1) if err_match else err
-        raise RuntimeError(errmessage)
+        pattern = r"Error: \[NodeId: (?P<id>[a-zA-Z0-9-_]*)\] " \
+                  r"Operator \'[a-zA-Z0-9-_]*\': " \
+                  r"Unknown element \'(?P<par>[a-zA-Z]*)\'"
+        match = re.search(pattern, err)
+        if match is not None:
+            replace = match.groupdict()
+            with Workflow(xmlfile) as flow:
+                print('  removing parameter {id}:{par} and executing modified workflow'.format(**replace))
+                node = flow[replace['id']]
+                del node.parameters[replace['par']]
+                flow.write(xmlfile)
+            execute(xmlfile, cleanup=cleanup, gpt_exceptions=gpt_exceptions, verbose=False)
+        else:
+            if cleanup:
+                if os.path.isfile(outname + '.tif'):
+                    os.remove(outname + '.tif')
+                elif os.path.isdir(outname):
+                    shutil.rmtree(outname)
+            print(out + err)
+            print('failed: {}'.format(os.path.basename(infile)))
+            err_match = re.search('Error: (.*)\n', out + err)
+            errmessage = err_match.group(1) if err_match else err
+            raise RuntimeError(errmessage)
 
 
 def gpt(xmlfile, groups=None, cleanup=True, gpt_exceptions=None):
@@ -588,6 +604,7 @@ class Workflow(object):
     xmlfile: str
         the workflow XML file
     """
+    
     def __init__(self, xmlfile):
         with open(xmlfile, 'r') as infile:
             self.tree = ET.fromstring(infile.read())
@@ -702,7 +719,7 @@ class Workflow(object):
                 # print('resetting source of successor {} to {}'.format(successor.id, newnode.id))
                 successor.source = newnode.id
             # else:
-                # print('no source resetting required')
+            # print('no source resetting required')
         else:
             self.tree.insert(len(self.tree) - 1, node.element)
         self.refresh_ids()
@@ -730,7 +747,7 @@ class Workflow(object):
             else:
                 counter[operator] += 1
                 node.id = '{} ({})'.format(operator, counter[operator])
-                
+    
     @property
     def suffix(self):
         """
@@ -771,6 +788,7 @@ class Node(object):
     element: ~xml.etree.ElementTree.Element
         the node XML element
     """
+    
     def __init__(self, element):
         if not isinstance(element, ET.Element):
             raise TypeError('element must be of type xml.etree.ElementTree.Element')
@@ -851,6 +869,7 @@ class Par(object):
     element: ~xml.etree.ElementTree.Element
         the node parameter XML element
     """
+    
     def __init__(self, element):
         self.__element = element
